rework slash commands for further future compatablility

+++ src/discord/Commands/LinkMe.js
@@ -5,7 +5,7 @@ const Embeds = require("../Embeds");
 const mojangRequest = require("../../mojangRequest");
 const utils = require("../../utils");
 
-module.exports = new Command("linkme", ["*"], async (args, rawMsg) => {
+module.exports = new Command("linkme", ["*"], async (args, rawMsg, interaction) => {
     let player = args[0];
     if (player == undefined) {
         let embed = Embeds.errIptIgn;
@@ -20,11 +20,18 @@ module.exports = new Command("linkme", ["*"], async (args, rawMsg) => {
     if (acc == undefined) {
         let embed = Embeds.waiting;
 
-        let tmpMsg = await rawMsg.channel.send("", { embed: embed });
+        let tmpMsg;
+        if(interaction == undefined) {
+            tmpMsg = await rawMsg.channel.send("", { embed: embed });
+        } else {
+            interaction.defer();
+        }
         let uuid =
             player.length == 32 ? player : await mojangRequest.getUUID(player);
         if (("" + uuid).length != 32) {
-            await tmpMsg.delete();
+            if(interaction == undefined) {
+                await tmpMsg.delete();
+            }
             let noexistEmbed = Embeds.errIgnNull;
 
             return { res: "", embed: noexistEmbed };
@@ -32,19 +39,30 @@ module.exports = new Command("linkme", ["*"], async (args, rawMsg) => {
         acc = new Account(player, 0, uuid);
         await addAccounts("others", [uuid]);
         await acc.updateHypixel();
-        await tmpMsg.delete();
+        if(interaction == undefined) {
+            await tmpMsg.delete();
+        }
     }
 
-    if (
-        ("" + acc.hypixelDiscord).toLowerCase() ==
-        rawMsg.author.tag.toLowerCase()
-    ) {
+    let tag;
+    if(interaction == undefined) {
+        tag = rawMsg.author.tag.toLowerCase();
+    } else {
+        tag = interaction.member.user.tag.tpLowerCase();
+    }
+
+    if (("" + acc.hypixelDiscord).toLowerCase() == tag) {
         let uuid = player;
         // if its not a uuid then convert to uuid
         if (player.length < 17) {
             uuid = acc.uuid;
         }
-        let discord = rawMsg.author.id;
+        let discord;
+        if(interaction == undefined) {
+            discord = rawMsg.author.id;
+        } else {
+            discord = interaction.member.id;
+        }
         let disclist = await utils.readJSON("./disclist.json");
         // make sure player isnt linked
         if (disclist[discord]) {
